First commit for the new design
This commit contains the FINAL2.pdf which a member of the section has made (Ajla) as a separate project. This PDF contains
the design which the final design for the new website for dtek is recommended to build upon.

I have committed a beginning of the design. It uses node to use webpack to put together the files in a nice way.

Use `npm install` then `npm run build` to build into dist.